updated integers to float for better average

+++ db/migrate/20121004175007_add_value_to_ratings.rb
@@ -1,5 +1,5 @@
 class AddValueToRatings < ActiveRecord::Migration
   def change
-    add_column :ratings, :value, :integer
+    add_column :ratings, :value, :float
   end
 end
