+++ db/migrate/20121012165711_add_rating_total_to_photo.rb
@@ -1,5 +1,5 @@
 class AddRatingsTotalToPhoto < ActiveRecord::Migration
   def change
-  	add_column :photos, :ratings_total, :integer
+  	add_column :photos, :ratings_total, :float
   end
 end
