+++ db/migrate/20121013225648_add_ratings_count_to_photo.rb
@@ -1,5 +1,5 @@
 class AddRatingsCountToPhoto < ActiveRecord::Migration
   def change
-  	add_column :photos, :ratings_count, :integer
+  	add_column :photos, :ratings_count, :float
   end
 end
