+++ db/migrate/20121015032447_add_ratings_ave_to_photo.rb
@@ -1,5 +1,5 @@
 class AddRatingsAveToPhoto < ActiveRecord::Migration
   def change
-  	add_column :photos, :ratings_ave, :integer
+  	add_column :photos, :ratings_ave, :float
   end
 end
